+++ src/main/java/br/com/zalf/prolog/webservice/gente/solicitacaoFolga/relatorios/SolicitacaoFolgaRelatorioResource.java
@@ -1,0 +1,38 @@
+package br.com.zalf.prolog.webservice.gente.solicitacaoFolga.relatorios;
+
+import br.com.zalf.prolog.commons.Report;
+import br.com.zalf.prolog.webservice.interceptors.auth.Secured;
+
+import javax.ws.rs.*;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.StreamingOutput;
+
+/**
+ * Created by Zart on 02/05/17.
+ */
+@Consumes(MediaType.APPLICATION_JSON + ";charset=utf-8")
+@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
+@Path("/solicitacao-folgas/relatorios")
+public class SolicitacaoFolgaRelatorioResource {
+
+    SolicitacaoFolgaRelatorioService service = new SolicitacaoFolgaRelatorioService();
+
+    @GET
+    @Path("/resumos/{codUnidade}/csv")
+    @Secured
+    public StreamingOutput getResumoFolgasConcedidasCsv(@PathParam("codUnidade") Long codUnidade,
+                                                        @QueryParam("dataInicial") long dataInicial,
+                                                        @QueryParam("dataFinal") long dataFinal) {
+        return outputStream -> service.getResumoFolgasConcedidasCsv(codUnidade, outputStream, dataInicial, dataFinal);
+    }
+
+    @GET
+    @Secured
+    @Path("/resumos/{codUnidade}/report")
+    public Report getResumoFolgasConcedidasReport(@PathParam("codUnidade") Long codUnidade,
+                                                  @QueryParam("dataInicial") long dataInicial,
+                                                  @QueryParam("dataFinal") long dataFinal) {
+        return service.getResumoFolgasConcedidasReport(codUnidade, dataInicial, dataFinal);
+    }
+
+}